requirements.txt
alabaster==0.7.10
amqp==2.2.2
ansible==2.3.2.0
asgi-rabbitmq==0.5.4
asgiref==1.1.2
asn1crypto==0.23.0
attrs==17.2.0
autobahn==17.10.1
Automat==0.6.0
Babel==2.5.1
bcrypt==3.1.4
billiard==3.5.0.3
cached-property==1.3.1
celery==4.1.0
certifi==2017.7.27.1
cffi==1.11.2
channels==1.1.8
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.4.1
cryptography==2.1.2
Django==1.11.6
django-celery-beat==1.1.0
django-celery-results==1.0.1
django-extensions==1.9.7
django-rest-swagger==2.1.2
django-select2-forms==2.0.2
django-sortedm2m==1.5.0
djangorestframework==3.7.1
docutils==0.14
ephem==3.7.6.0
flake8==3.5.0
hyperlink==17.3.1
idna==2.6
imagesize==0.7.1
incremental==17.5.0
itypes==1.1.0
Jinja2==2.9.6
kombu==4.1.0
MarkupSafe==1.0
mccabe==0.6.1
msgpack-python==0.4.8
openapi-codec==1.3.2
paramiko==2.3.1
pika==0.11.0
psycopg2==2.7.3.2
pushbullet.py==0.11.0
pyasn1==0.3.7
pycodestyle==2.3.1
pycparser==2.18
pycrypto==2.6.1
pyflakes==1.6.0
Pygments==2.2.0
PyNaCl==1.1.2
python-magic==0.4.13
pytz==2017.3
PyYAML==3.12
requests==2.18.4
simplejson==3.11.1
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.6.5
sphinx-rtd-theme==0.2.4
sphinxcontrib-websupport==1.0.1
Twisted==17.9.0
txaio==2.8.2
typing==3.6.2
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
websocket-client==0.44.0
zope.interface==4.4.3
